clean up all the things; make Sniffer class on server side
