+++ server/src/routes.py
@@ -1,3 +1,4 @@
+import sniff_module
 import aiohttp
 import json
 from aiohttp import web
@@ -5,7 +6,7 @@ from scapy.all import *
 import threading
 import asyncio
 import utils
-from utils import get_interface, start_dequeue_thread, start_sniff_thread, get_arp_table, arp_spoof
+from utils import get_interface, get_arp_table, arp_spoof
 
 def setup_routes(app):
     app.router.add_get('/sniff', sniff_handler),
@@ -50,12 +51,12 @@ async def sniff_handler(request):
                     print('Starting neighbor sniff')
                     neighbor_ip = json_msg['args']['ifaddr']
                     arp_spoof(neighbor_ip)
-                    start_sniff_thread(ws, json_msg['fname'], neighbor_ip)
-                    start_dequeue_thread(ws)
+                    s = sniff_module.Sniffer(ws, json_msg['fname'], neighbor_ip)
+                    s.start_sniff_threads()
                 elif json_msg['fname'] == 'sniffSelf':
                     print('Starting self sniff')
-                    start_sniff_thread(ws, json_msg['fname'])
-                    start_dequeue_thread(ws)
+                    s = sniff_module.Sniffer(ws, json_msg['fname'])
+                    s.start_sniff_threads()
             elif msg.type == aiohttp.WSMsgType.ERROR:
                 print('ws connection closed with exception %s' % ws.exception())
     except Exception as e:
